Create BCandEEPROM
Sends a prompt message to UART0 for input. That input is then stored in the EEPROM and read back. Then it is transmitted to the other board through UART1. Any input received through UART1 is stored on the EEPROM and read back. It is then sent to UART0 to be displayed.